+++ test/src/test/java/com/devexperts/timetest/test/TestTimeProviderTest.java
@@ -7,18 +7,18 @@ package com.devexperts.timetest.test;
  * Copyright (C) 2015 - 2016 Devexperts, LLC
  * %%
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
+ * it under the terms of the GNU Lesser General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU General Lesser Public License for more details.
  * 
- * You should have received a copy of the GNU General Public
+ * You should have received a copy of the GNU General Lesser Public
  * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-3.0.html>.
+ * <http://www.gnu.org/licenses/lgpl-3.0.html>.
  * #L%
  */
 
@@ -27,6 +27,8 @@ import com.devexperts.timetest.TestTimeProvider;
 import com.devexperts.util.UnsafeHolder;
 import org.junit.After;
 import org.junit.Test;
+
+import java.util.concurrent.locks.LockSupport;
 
 import static org.junit.Assert.*;
 
@@ -260,6 +262,27 @@ public class TestTimeProviderTest {
         TestTimeProvider.waitUntilThreadsAreFrozen(100);
         synchronized (owner) {
             UnsafeHolder.UNSAFE.unpark(thread);
+            owner.wait(Long.MAX_VALUE);
+        }
+    }
+
+    @Test(timeout = 1000)
+    public void testLockSupport() throws InterruptedException {
+        TestTimeProvider.start();
+        final Object owner = new Object();
+        Thread thread = new Thread(new Runnable() {
+            @Override
+            public void run() {
+                LockSupport.park();
+                synchronized (owner) {
+                    owner.notifyAll();
+                }
+            }
+        });
+        thread.start();
+        TestTimeProvider.waitUntilThreadsAreFrozen(100);
+        synchronized (owner) {
+            LockSupport.unpark(thread);
             owner.wait(Long.MAX_VALUE);
         }
     }