+++ transformer/src/main/java/com/devexperts/timetest/ChangeTimeMethodsTransformer.java
@@ -7,24 +7,24 @@ package com.devexperts.timetest;
  * Copyright (C) 2015 - 2016 Devexperts, LLC
  * %%
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
+ * it under the terms of the GNU Lesser General Public License as
  * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU General Lesser Public License for more details.
  * 
- * You should have received a copy of the GNU General Public
+ * You should have received a copy of the GNU General Lesser Public
  * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-3.0.html>.
+ * <http://www.gnu.org/licenses/lgpl-3.0.html>.
  * #L%
  */
 
+import com.devexperts.jagent.*;
 import org.objectweb.asm.*;
 
-import java.lang.instrument.ClassFileTransformer;
 import java.lang.instrument.IllegalClassFormatException;
 import java.security.ProtectionDomain;
 import java.util.ArrayList;
@@ -33,13 +33,16 @@ import java.util.regex.Pattern;
 
 import static org.objectweb.asm.Opcodes.*;
 
-public class ChangeTimeMethodsTransformer implements ClassFileTransformer {
+public class ChangeTimeMethodsTransformer extends CachingClassFileTransformer {
 
     private static final int ASM_API = Opcodes.ASM5;
     private final List<Pattern> includes;
     private final List<Pattern> excludes;
+    private final ClassInfoCache ciCache;
 
-    public ChangeTimeMethodsTransformer(Configuration configuration) {
+    public ChangeTimeMethodsTransformer(Configuration configuration, Log log, String agentVersion) {
+        super(log, agentVersion);
+        this.ciCache = new ClassInfoCache(log);
         includes = new ArrayList<>(configuration.include().length);
         for (String s : configuration.include())
             includes.add(GlobUtil.compile(s.replaceAll("\\.", "/")));
@@ -49,13 +52,14 @@ public class ChangeTimeMethodsTransformer implements ClassFileTransformer {
     }
 
     @Override
-    public byte[] transform(ClassLoader loader, final String className, Class<?> classBeingRedefined, ProtectionDomain protectionDomain,
-                            byte[] classfileBuffer) throws IllegalClassFormatException {
-        if ((className.startsWith("com/devexperts/timetest/") && !className.startsWith("com/devexperts/timetest/test/")) ||
-                className.startsWith("sun/") ||
-                className.startsWith("java/") ||
-                className.startsWith("org/apache/maven/"))
-            return null;
+    protected boolean processClass(String className, ClassLoader loader) {
+        if ((className.startsWith("com/devexperts/timetest/") && !className.startsWith("com/devexperts/timetest/test/"))
+                || className.startsWith("com/sun/")
+                || className.startsWith("sun/")
+                || className.startsWith("jdk/")
+                || className.startsWith("javax/")
+                || (className.startsWith("java/") && !className.startsWith("java/util/concurrent/")))
+            return false;
 
         boolean process = false;
         for (Pattern p : includes) {
@@ -70,11 +74,17 @@ public class ChangeTimeMethodsTransformer implements ClassFileTransformer {
                 break;
             }
         }
-        if (!process)
-            return null;
+        return process;
+    }
 
+    @Override
+    public byte[] transformImpl(ClassLoader loader, final String className, Class<?> classBeingRedefined, ProtectionDomain protectionDomain,
+                            byte[] classfileBuffer) throws IllegalClassFormatException {
         ClassReader cr = new ClassReader(classfileBuffer);
-        ClassWriter cw = new ClassWriter(ClassWriter.COMPUTE_MAXS);
+        ClassInfoVisitor ciVisitor = new ClassInfoVisitor();
+        cr.accept(ciVisitor, 0);
+        ciCache.getOrInitClassInfoMap(loader).put(className, ciVisitor.getClassInfo());
+        ClassWriter cw = new FrameClassWriter(loader, ciCache);
         ClassVisitor transformer = new ClassVisitor(ASM_API, cw) {
             @Override
             public MethodVisitor visitMethod(int access, final String mname, String mdesc, String signature, String[] exceptions) {