+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 Changelog
 =========
 
+1.1 - 2016-10-27
+----------------
+* Support class caching
+* Fix bug with frames processing in ASM
+* Change license to LGPLv3
+* shade dxlib into agent
+
+1.0 - 2016-06-27
+----------------
+* Migrate to JAgent with OWNER configuration.
+
 0.9 - 2015-10-13
 ----------------
-
-Initial public version
+* Initial public version